requirements.txt
Flask==2.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.6
blinker==1.9.0
requests==2.31.0
psycopg2-binary==2.9.9
